requirements.txt
scipy>=0.18.1
numpy>=1.11.2
pytest==3.2.5
biopython
